requirements.txt
# Web and API
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0

# AI/ML
openai>=1.0.0

# Optional (if using Replicate)
replicate>=0.20.0
